requirements.txt
prettytable == 0.7.2
requests == 2.7.0
ansicolors == 1.0.2
